Fix up documentation breakages from moving to plot sub-package
