+++ README.md
@@ -75,7 +75,7 @@ additional functionality for writing analysis programs.
 -   **Stoner.Analysis.AnalysisMixin** provides additional methods for curve-fitting, differentiating, smoothing and carrying out
     :   basic calculations on data.
 
--   **Stoner.Plot.PlotMixin** provides additional routines for plotting
+-   **Stoner.plot.PlotMixin** provides additional routines for plotting
     data on 2D or 3D plots.
 
 For rapid development of small scripts, we would recommend the
