+++ doc/Stoner.rst
@@ -36,19 +36,19 @@ Functions
    :functions-only:
 
 
-:mod:`Plot` Module
-==================
+:mod:`plot` Package
+====================
 
 Classes
 -------
-.. automodsumm:: Stoner.Plot
+.. automodsumm:: Stoner.plot.core
    :toctree: classes
    :template: classdocs.rst
    :classes-only:
 
 Functions
 ---------
-.. automodsumm:: Stoner.Plot
+.. automodsumm:: Stoner.plot.core
    :toctree: functions
    :template: funcdocs.rst
    :functions-only:
@@ -116,34 +116,19 @@ Facility Outputs
 
 Classes
 -------
-.. automodsumm:: Stoner.PlotFormats
+.. automodsumm:: Stoner.plot.formats
    :toctree: classes
    :template: classdocs.rst
    :classes-only:
-
-Functions
----------
-.. automodsumm:: Stoner.PlotFormats
-   :toctree: functions
-   :template: funcdocs.rst
-   :functions-only:
 
 See also the stylesheets available in the stylelib directory.
 
-:mod:`plotutils` Module
+:mod:`plot.utils` Module
 =======================
-
-
-Classes
--------
-.. automodsumm:: Stoner.plotutils
-   :toctree: classes
-   :template: classdocs.rst
-   :classes-only:
 
 Functions
 ---------
-.. automodsumm:: Stoner.plotutils
+.. automodsumm:: Stoner.plot.utils
    :toctree: functions
    :template: funcdocs.rst
    :functions-only:
