+++ doc/UserGuide/cookbook.rst
@@ -17,7 +17,7 @@ The :py:class:`Data` Class
 --------------------------
 
 In practice, one often requires both the analysis functions of :py:class:`Stoner.Analysis.AnalysisMixin` and
-the plotting functions of :py:class:`Stoner.Plot.PlotFile`. This can be done easily by creating a
+the plotting functions of :py:class:`Stoner.plot.PlotMixin`. This can be done easily by creating a
 subclass that inherits from both, but for convenience, the :py:mod:`Stoner.Util` module provides the :py:class:`Data`
 class that does this for you.::
 
@@ -168,8 +168,8 @@ data. After the **for** loop we set the column headers in *t* and then insert
 an additional column at the start with the gate voltage values.
 
 The matrix generated by this code is suitable for feeding directly into
-:py:meth:`Stoner.Plot.PlotMixin.plot_matrix`, however, the same plot could be generated
-directly from the :py:meth:`Stoner.Plot.PlotMixin.plot_xyz` method too.
+:py:meth:`Stoner.plot.PlotMixin.plot_matrix`, however, the same plot could be generated
+directly from the :py:meth:`Stoner.plot.PlotMixin.plot_xyz` method too.
 
 Sampling a A 3D Surface function
 --------------------------------
@@ -179,12 +179,12 @@ Suppose you have a function Z(x,y) that is defined over a range -X0..+X0 and
 by creating a random distribution of (x,y) points, evaluating the function and then
 using the :py:class:`PlotFile` methods to generate the surface.::
 
-    from Stoner.Plot import PlotFile
+    from Stoner import Data
     from numpy import uniform
 
     samples=5000
 
-    surf=PlotFile()
+    surf=Data()
     surf&=uniform(-X0,X0,samples)
     surf&=uniform(-Y0,Y0,samples)
     surf.setas="xy"
@@ -201,7 +201,7 @@ be able to examine a cross-section of the simulation. The Stoner package provide
 to quickly examine the output data::
 
     from Stoner.FileFormats import OVFFile # reads OOMMF vector field files
-    import Stoner.Plot
+    import Stoner.plot
     p=SP.PlotFile('my_simulation.ovf')
     p.setas="xyzuvw"
     p=p.section(z=10.5) # Take a slice inthe xy plane where z is 10.5 nm
