+++ doc/samples/plotstyles/GBStyle.py
@@ -1,5 +1,5 @@
 from Stoner import Data
-from Stoner.PlotFormats import GBPlotStyle
+from Stoner.plot.formats import GBPlotStyle
 
 d=Data("../sample.txt",setas="xy",template=GBPlotStyle)
 d.y=d.y-(max(d.y)/2)
