+++ doc/samples/plotstyles/PRBStyle.py
@@ -1,5 +1,5 @@
 from Stoner import Data
-from Stoner.PlotFormats import PRBPlotStyle
+from Stoner.plot.formats import PRBPlotStyle
 
 d=Data("../sample.txt",setas="xy",template=PRBPlotStyle)
 d.plot()
