+++ doc/samples/plotstyles/SeabornStyle.py
@@ -1,5 +1,5 @@
 from Stoner import Data
-from Stoner.PlotFormats import SeabornPlotStyle
+from Stoner.plot.formats import SeabornPlotStyle
 
 d=Data("../sample.txt",setas="xyy",template=SeabornPlotStyle(stylename="dark",context="talk",palette="muted"))
 d.plot(multiple="y2")
