+++ doc/samples/plotstyles/SketchStyle.py
@@ -1,5 +1,5 @@
 from Stoner import Data
-from Stoner.PlotFormats import SketchPlot    
+from Stoner.plot.formats import SketchPlot    
 
 d=Data("../sample.txt",setas="xy",template=SketchPlot)
 d.plot()
