+++ doc/samples/plotstyles/default.py
@@ -1,5 +1,5 @@
 from Stoner import Data
-from Stoner.PlotFormats import DefaultPlotStyle
+from Stoner.plot.formats import DefaultPlotStyle
 
 d=Data("../sample.txt",setas="xy",template=DefaultPlotStyle)
 d.plot()
