+++ doc/samples/template.py
@@ -1,5 +1,5 @@
 from Stoner import Data
-from Stoner.PlotFormats import DefaultPlotStyle
+from Stoner.plot.formats import DefaultPlotStyle
 p=Data("sample.txt",setas="xy",template=DefaultPlotStyle())
 p.template(axes_color__cycle=["r","g","b"])
 p.plot()
