+++ doc/samples/template2.py
@@ -1,4 +1,4 @@
 from Stoner import Data
-from Stoner.PlotFormats import SketchPlot
+from Stoner.plot.formats import SketchPlot
 p=Data("sample.txt",setas="xy",template=SketchPlot)
 p.plot()
